pyproject.toml
[tool.poetry]
name = "diploma"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
django = "^4.1.5"
djangorestframework = "^3.14.0"
psycopg2-binary = "^2.9.5"
envparse = "^0.2.0"
social-auth-app-django = "^5.0.0"
django-cors-headers = "^3.13.0"
django-filter = "^22.1"
marshmallow = "^3.19.0"
marshmallow-dataclass = "^8.5.11"
requests = "^2.28.2"
drf-spectacular = "^0.25.1"
pytest-django = "^4.5.2"
pytest-factoryboy = "^2.5.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
